pyproject.toml
[project]
name = "sc-tools"
version = "0.1.0"
description = "Single-cell tools for analysis and visualization"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anndata>=0.10",
    "geoparse>=2.0",
    "gdown>=5.1",
    "goatools>=1.4",
    "gradio>=4.0",
    "harmonypy>=0.0.10",
    "huggingface-hub>=0.21",
    "matplotlib>=3.8",
    "numpy>=1.23",
    "omegaconf>=2.3",
    "openai>=1.0",
    "packaging>=23.0",
    "pandas>=1.5",
    "plotly>=6.6.0",
    "pronto>=2.5",
    "python-dotenv>=1.0",
    "rapidfuzz>=3.0",
    "requests>=2.31",
    "rich>=13.0",
    "scanorama>=1.7.4",
    "scanpy>=1.9",
    "scikit-learn>=1.2",
    "scvi-tools>=1.1",
    "seaborn>=0.13",
    "torch>=2.11.0",
    "tqdm>=4.65",
    "transformers>=4.36",
    "umap-learn>=0.5",
    "wandb>=0.16",
]